+++ R/textplot.R
@@ -32,14 +32,21 @@ textplot_simil.textmodel_lss <- function(x) {
 
 #' Plot polarity scores of words
 #' @param x a fitted textmodel_lss object.
-#' @param highlighted [quanteda::pattern] to select words to highlight.
+#' @param highlighted [quanteda::pattern] to select words to highlight. If a
+#'   [quanteda::dictionary] is passed, words in the top-level categories are
+#'   highlighted in different colors.
 #' @param max_highlighted the maximum number of words to highlight. When
 #'   `highlighted = NULL`, words to highlight are randomly selected
 #'   proportionally to `polarity ^ 2 * log(frequency)`.
 #' @param max_words the maximum number of words to plot. Words are randomly
 #'   sampled to keep the number below the limit.
-#' @param ... passed to [ggplot2::geom_text()] and
-#'   [ggrepel::geom_text_repel()] to customize text labels.
+#' @param ... passed to underlying functions. See the Details.
+#' @details Users can customize the plots through `...`, which is
+#'   passed to [ggplot2::geom_text()] and [ggrepel::geom_text_repel()]. The
+#'   colors are specified internally but users can override the settings by appending
+#'   [ggplot2::scale_colour_manual()] or [ggplot2::scale_colour_brewer()]. The
+#'   legend title can also be modified using [ggplot2::labs()].
+#' @importFrom ggrepel geom_text_repel
 #' @export
 textplot_terms <- function(x, highlighted = NULL,
                            max_highlighted = 50, max_words = 1000, ...) {
@@ -66,50 +73,67 @@ textplot_terms.textmodel_lss <- function(x, highlighted = NULL,
 
     temp <- subset(temp, freq > 0)
     temp$freq <- log(temp$freq)
+    temp$id <- seq_len(nrow(temp))
+    temp$group <- factor(rep("highlighted", nrow(temp)))
 
     if (is.null(highlighted)) {
-        id <- seq_len(nrow(temp))
+        key <- NULL
     } else {
         if (is.dictionary(highlighted)) {
             separator <- meta(highlighted, field = "separator", type = "object")
             valuetype <- meta(highlighted, field = "valuetype", type = "object")
-            concatenator <- x$concatenator
-            highlighted <- unlist(highlighted, use.names = FALSE)
-            if (!nzchar(separator) && !is.null(concatenator)) # for backward compatibility
-                highlighted <- stri_replace_all_fixed(highlighted, separator, concatenator)
         } else {
             highlighted <- unlist(highlighted, use.names = FALSE)
             valuetype <- "glob"
         }
-
-        id <- unlist(quanteda::pattern2id(
+        ids <- quanteda::object2id(
             highlighted,
-            types = names(x$beta),
+            types = temp$word,
             valuetype = valuetype,
-            case_insensitive = TRUE
-        ), use.names = FALSE)
+            case_insensitive = TRUE,
+            concatenator = x$concatenator
+        )
+        key <- attr(ids, "key")
+        id <- unlist(ids)
+        if (!is.null(key) && !is.null(id)) {
+            temp$group <- factor(names(id[match(temp$id, id)]), levels = key)
+        } else {
+            temp$group <- factor(ifelse(temp$id %in% id, "highlighted", NA),
+                                 levels = "highlighted")
+        }
     }
-    i <- seq_len(nrow(temp))
-    p <- as.numeric(i %in% id) * temp$beta ^ 2 * temp$freq
-    if (all(p == 0)) {
-        l <- rep(FALSE, length(i))
+    temp$p <- as.numeric(!is.na(temp$group)) * temp$beta ^ 2 * temp$freq
+    if (all(temp$p == 0)) {
+        l <- rep(FALSE, length(temp$id))
     } else {
-        l <- i %in% sample(i, min(sum(p > 0), max_highlighted), prob = p)
+        l <- temp$id %in% sample(temp$id, min(sum(temp$p > 0), max_highlighted),
+                                 prob = temp$p)
     }
     temp_hi <- temp[l,]
     temp_lo <- temp[!l,]
 
     temp_lo <- head(temp_lo[sample(seq_len(nrow(temp_lo))),], max_words)
-    ggplot(data = temp_lo, aes(x = beta, y = freq, label = word)) +
+    gg <- ggplot(data = temp_lo, aes(x = beta, y = freq, label = word)) +
            geom_text(colour = "grey70", alpha = 0.7, ...) +
            labs(x = "Polarity", y = "Frequency (log)") +
            theme_bw() +
            theme(panel.grid= element_blank(),
                  axis.title.x = element_text(margin = margin(t = 10, r = 0, b = 0, l = 0)),
                  axis.title.y = element_text(margin = margin(t = 0, r = 10, b = 0, l = 0))) +
-           geom_text_repel(data = temp_hi, aes(x = beta, y = freq, label = word),
-                           segment.size = 0.25, colour = "black", ...) +
-           geom_point(data = temp_hi, aes(x = beta, y = freq), cex = 0.7, colour = "black")
+           geom_text_repel(data = temp_hi, aes(x = beta, y = freq, label = word, colour = group),
+                           segment.size = 0.25, show.legend = FALSE, ...) +
+           geom_point(data = temp_hi, aes(x = beta, y = freq, shape = group, colour = group),
+                      cex = 1)
+
+    if (!is.null(key) && length(key) > 1) {
+        gg <- gg + scale_colour_brewer(palette = "Set1", drop = FALSE) +
+                   scale_shape_discrete(drop = FALSE) +
+                   labs(colour = "Group", shape = "Group")
+    } else {
+        gg <- gg + scale_colour_manual(values = "black") +
+                   guides(colour = "none", shape = "none")
+    }
+    return(gg)
 }
 
 #' \[experimental\] Plot clusters of word vectors